requirements.txt
apache-airflow==2.5.0
mlflow==2.7.0
pandas==1.5.3
scikit-learn==1.2.2